Added Get request for the dust data
