+++ fullstack/src/app/api/recieve-data/route.js
@@ -1,10 +1,23 @@
+/**
+ * @file This file is the route for the recieve-data API
+ * @module app/api/recieve-data
+ * @description This file is the route for the recieve-data API
+ * - The POST method is used to recieve data from the webhook
+ * - The data is then seperated into different tables based on the data type
+ * - The data is then sent to the database which have been functioned to keep code readable
+ * and will be found in the receive-data-helper.js utils file
+ */
 import { DustData } from "@/app/utils/recieve-data-helper";
-import db from "@/db";
 
 export const POST = async (request) => {
     try {
         const data = await request.json();
         let send = []
+
+        /**
+         * Since the webhook will send all data we will seperate the data
+         * into different tables based on there data type
+         */
         const device_id = data.uplink_message.f_port;
         const humidity = data.uplink_message.decoded_payload.humidity;
         const temperature = data.uplink_message.decoded_payload.temperature;
@@ -31,18 +44,5 @@ export const POST = async (request) => {
             headers: { 'Content-Type': 'application/json' },
             status: 500
         });
-    }
-}
-
-export const GET = async () => {
-    try {
-        const data = await db.any('SELECT * FROM sensor_pressure');
-
-        return new Response(JSON.stringify(data), {
-            status: 200
-        });
-    } catch (err) {
-        console.error(err);
-        response.status(500).send('Server Error');
     }
 }
